+++ docs/versionhistory.rst
@@ -3,6 +3,19 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
+**UNRELEASED**
+
+- Added 4 new fixtures for the AnyIO ``pytest`` plugin:
+
+  * ``free_tcp_port_factory``: session scoped fixture returning a callable that
+    generates unused TCP port numbers
+  * ``free_udp_port_factory``: session scoped fixture returning a callable that
+    generates unused UDP port numbers
+  * ``free_tcp_port``: function level fixture that invokes the ``free_tcp_port_factory``
+    fixture to generate a free TCP port number
+  * ``free_udp_port``: function level fixture that invokes the ``free_udp_port_factory``
+    fixture to generate a free UDP port number
+
 **4.8.0**
 
 - Added **experimental** support for running functions in subinterpreters on Python